--- src/app/services/database.service.spec.ts
@@ -1,12 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { DatabaseService } from './database.service';
-
-describe('DatabaseService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
-
-  it('should be created', () => {
-    const service: DatabaseService = TestBed.get(DatabaseService);
-    expect(service).toBeTruthy();
-  });
-});
